moved all unit tests to a separate folder
